+++ src/modules/mod.rs
@@ -1,7 +1,7 @@
 pub mod modem;
 pub mod sensors;
-pub mod tasks;
 pub mod util;
 pub mod config;
 pub mod network;
 pub mod bus;
+pub mod watchdog;
