+++ src/modules/sensors/mod.rs
@@ -1,3 +1,2 @@
 pub mod gas;
-pub mod types;
 pub mod battery;
