--- src/modules/tasks/mod.rs
@@ -1,2 +1,0 @@
-pub mod system;
-pub mod timing;
